+++ src/theme/separator.ts
@@ -46,6 +46,11 @@ export default {
       thin: '',
       thick: ''
     },
+    position: {
+      start: '',
+      center: '',
+      end: ''
+    },
     type: {
       solid: {
         border: 'border-solid'
@@ -62,6 +67,38 @@ export default {
     }
   },
   compoundVariants: [
+    // region position ////
+    {
+      orientation: 'horizontal',
+      position: 'start',
+      class: { container: 'me-3' }
+    },
+    {
+      orientation: 'horizontal',
+      position: 'center',
+      class: { container: 'mx-3' }
+    },
+    {
+      orientation: 'horizontal',
+      position: 'end',
+      class: { container: 'ms-3' }
+    },
+    {
+      orientation: 'vertical',
+      position: 'start',
+      class: { container: 'mb-2' }
+    },
+    {
+      orientation: 'vertical',
+      position: 'center',
+      class: { container: 'my-2' }
+    },
+    {
+      orientation: 'vertical',
+      position: 'end',
+      class: { container: 'mt-2' }
+    },
+    // endregion ////
     // region horizontal ////
     {
       orientation: 'horizontal',